New utils method to convert a given task argument into a bool
The goal of this method is to convert a fabric argument intended as a
bool representation into a Python bool.

In case that the argument is not a supported string representation (for
strtobool) of a bool it will raise an exception.

If the argument is not a string it will return a cast of the argument
into a bool.